Move contents of header to source of DTConfigDBProducer
